requirements.txt
Flask
Flask-RESTful
flask_cors
flask_injector
passlib
azure-cosmos